requirements.txt
fastapi
uvicorn
openai
langchain
python-dotenv
pydantic
beautifulsoup4
